ecommerce_pipeline/requirements.txt
# E-Commerce Clickstream Pipeline - Python Dependencies
# Core: PySpark, Kafka, Pydantic
# Testing: pytest

# Core Data Processing
pyspark==3.5.0
pandas==2.1.0
pyarrow==14.0.0

# Kafka Integration
confluent-kafka==2.3.0

# Configuration & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Airflow (for local development)
apache-airflow==2.8.0
apache-airflow-providers-apache-spark==4.5.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Utilities
python-json-logger==2.0.7
